fix Scala 2.11 compatibility issues
