+++ core/shared/src/main/scala/zio/Has.scala
@@ -16,8 +16,6 @@
 
 package zio
 
-import com.github.ghik.silencer.silent
-
 /**
  * The trait `Has[A]` is used with ZIO environment to express an effect's
  * dependency on a service of type `A`. For example,
@@ -55,18 +53,6 @@ object Has {
 
   type MustHave[A, B]    = A <:< Has[B]
   type MustNotHave[A, B] = NotExtends[A, Has[B]]
-
-  @silent("define classes/objects inside of package objects")
-  trait NotExtends[A, B] extends Serializable
-
-  object NotExtends {
-    implicit def notExtends0[A, B]: A NotExtends B      = new NotExtends[A, B] {}
-    implicit def notExtends1[A <: B, B]: A NotExtends B = ???
-    @annotation.implicitAmbiguous(
-      "The environment ${A} already contains service ${B}, are you sure you want to overwrite it? Use Has#update to update a service already inside the environment."
-    )
-    implicit def notExtends2[A <: B, B]: A NotExtends B = ???
-  }
 
   trait IsHas[-R] {
     def add[R0 <: R, M: Tagged](r: R0, m: M): R0 with Has[M]
